refactor(daemon): extract onboarding preflight surface
